chore: refactor Contract Management section name
